pyproject.toml
[project]
name = "crewai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-search-results>=2.4.2",
    "python-dotenv>=1.2.2",
]
